+++ packages/gen-typescript-declarations/src/test/goldens/paper-behaviors-3/paper-checked-element-behavior.d.ts
@@ -8,13 +8,9 @@
  *   paper-checked-element-behavior.js
  */
 
-import {IronCheckedElementBehaviorImpl} from '@polymer/iron-checked-element-behavior/iron-checked-element-behavior.js';
-
 import {PaperInkyFocusBehavior} from './paper-inky-focus-behavior.js';
 
 import {PaperRippleBehavior} from './paper-ripple-behavior.js';
-
-export {PaperCheckedElementBehaviorImpl};
 
 /**
  * Use `PaperCheckedElementBehavior` to implement a custom element that has a