+++ packages/gen-typescript-declarations/src/test/goldens/paper-button-3/paper-button.d.ts
@@ -8,7 +8,7 @@
  *   paper-button.js
  */
 
-import {PaperButtonBehavior, PaperButtonBehaviorImpl} from '@polymer/paper-behaviors/paper-button-behavior.js';
+import {PaperButtonBehavior} from '@polymer/paper-behaviors/paper-button-behavior.js';
 
 import {Polymer} from '@polymer/polymer/lib/legacy/polymer-fn.js';